requirements.txt
Django==2.1.7
Pillow==5.4.1
pytz==2018.9
Cython==0.28.3
enum34==1.1.6
grpcio==1.19.0
grpcio-tools==1.19.0
protobuf==3.7.0
six==1.12.0
